+++ docs/release-notes.md
@@ -9,6 +9,7 @@
 
 ## Upcoming release
 * For mod authors:
+  * Added validation for the manifest `Dependencies` field.
   * Fixed validation for mods with version `0.0.0`.
 
 ## 3.10.1